fix(file-watcher): create specific watcher for module
Nuxt.js changes his way of exposing file watcher, so for better compatibility
we can use ower own file watcher
