+++ lib/constants.js
@@ -1,0 +1,17 @@
+
+export const DEFAULT_OPTIONS = {
+  routerNativeAlias: false
+}
+
+export const VALID_EXTRAS = [
+  'path',
+  'alias',
+  'meta',
+  'props'
+]
+
+export const NUXT_VALID_EXTRAS = [
+  'path',
+  'meta',
+  'props'
+]
