+++ lib/module.js
@@ -1,89 +1,49 @@
-import { pick, extractRouterBlockFromFile, parseYAML, validateExtras, NUXT_VALID_EXTRAS } from './utils'
+import path from 'path'
+import chokidar from 'chokidar'
 
-const DEFAULTS = {
-  routerNativeAlias: false
+import { extendRoutes, invalidateRoute } from './extras'
+import { DEFAULT_OPTIONS } from './constants'
+
+export default async function module(moduleOptions) {
+  const options = { ...DEFAULT_OPTIONS, ...moduleOptions, ...this.options.routerExtras }
+
+  await init.call(this, options)
+
+  await setupHooks.call(this, options)
 }
 
-const parsedRoutes = {}
-
-async function extractExtrasFromRoute(route, useCache = true) {
-  const previousExtras = parsedRoutes[route.component]
-  if (previousExtras && useCache) {
-    return previousExtras
-  }
-
-  let extras = {
-    _raw: '',
-    _needUpdate: false
-  }
-  const routerBlock = await extractRouterBlockFromFile(route.component)
-
-  if (routerBlock) {
-    const parsedData = parseYAML(routerBlock.content)
-    extras = validateExtras(parsedData)
-    // Keep raw value to detect changes
-    extras._raw = routerBlock.content
-    // Wheter route need to update or not
-    extras._needUpdate = !previousExtras || previousExtras._raw !== routerBlock.content
-  }
-  // Update cache
-  parsedRoutes[route.component] = extras
-
-  return extras
-}
-
-module.exports = function module(moduleOptions) {
-  const options = Object.assign({}, DEFAULTS, moduleOptions, this.options.routerExtras)
-  let localRouetes = []
-
-  this.nuxt.hook('build:done', (builder) => {
-    const filesWatcher = builder.watchers.files
-    if (filesWatcher) {
-      filesWatcher.on('change', async (file) => {
-        let changed = false
-
-        if (file.includes(builder.options.dir.pages)) {
-          const routes = localRouetes.filter(route => route.component === file)
-          if (routes.length) {
-            const extras = await extractExtrasFromRoute(routes[0], false)
-
-            if (extras._needUpdate) {
-              changed = true
-            }
-          }
-        }
-
-        if (changed) {
-          builder.generateRoutesAndFiles()
-        }
-      })
-    }
+function watchFiles(options, builder) {
+  const filesWatcher = options._filesWatcher = chokidar.watch(options.input, {
+    ignoreInitial: true
   })
 
-  this.nuxt.hook('build:extendRoutes', async (routes) => {
-    localRouetes = routes
-
-    const additionalRoutes = []
-
-    for (const route of routes) {
-      const extras = await extractExtrasFromRoute(route)
-
-      Object.assign(route, pick(extras, NUXT_VALID_EXTRAS))
-
-      if (options.routerNativeAlias) {
-        Object.assign(route, pick(extras, ['alias']))
-      } else if (extras.alias) {
-        for (const index in extras.alias) {
-          const alias = extras.alias[index]
-          additionalRoutes.push(Object.assign({}, route, {
-            path: alias,
-            name: `${route.name}-p${index}`
-          }))
-        }
+  if (filesWatcher) {
+    filesWatcher.on('change', async (file) => {
+      const isInvalid = await invalidateRoute(file)
+      if (isInvalid) {
+        builder.generateRoutesAndFiles()
       }
-    }
-    additionalRoutes.forEach(route => routes.push(route))
-  })
+    })
+  }
 }
 
-module.exports.meta = require('../package.json')
+function init(options) {
+  options.input = path.join(this.nuxt.options.srcDir, 'pages')
+
+  this.nuxt.hook('build:extendRoutes', extendRoutes)
+}
+
+function setupHooks(options) {
+  this.nuxt.hook('build:done', (builder) => {
+    if (this.nuxt.options.dev) {
+      watchFiles.call(this, options, builder)
+    }
+  })
+
+  this.nuxt.hook('close', () => {
+    if (options._filesWatcher) {
+      options._filesWatcher.close()
+      delete options._filesWatcher
+    }
+  })
+}
